+++ src/Menge/Layer2D.h
@@ -29,7 +29,7 @@ namespace Menge
 
 		bool handleKeyEvent( size_t _key, bool _isDown ) override;
 		bool handleMouseButtonEvent( size_t _button, bool _isDown ) override;
-		bool handleMouseMove( float _x, float _y, float _whell ) override;
+		bool handleMouseMove( int _x, int _y, int _whell ) override;
 
 	public:
 		bool _activate() override;
